refactor: modularize MetadataFetcher and add Zotero 9 compatibility
- Extract DOI discovery logic into DOIDiscoveryService
- Extract book/ISBN handling into BookMetadataService
- Extract metadata updates into MetadataUpdateService
- Create utility modules (isbn.ts, similarity.ts, itemFields.ts)
- Add dependency injection pattern for testable architecture
- Update version constraints for Zotero 9 (8.* -> 9.*)
- Optimize API call order with fallback strategy and rate limit delays
- Fix SemanticScholar rate limit config (100/min not 100/sec)
- Reduce MetadataFetcher from 2252 lines to ~1000 lines
