+++ addon/manifest.json
@@ -8,7 +8,7 @@
     "zotero": {
       "id": "zotadata@zotero.org",
       "strict_min_version": "8.0",
-      "strict_max_version": "8.*"
+      "strict_max_version": "9.*"
     }
   }
 }
