+++ src/__tests__/unit/semi-supervised-paper.test.ts
@@ -1,15 +1,74 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { MetadataFetcher } from "@/modules/MetadataFetcher";
+import {
+  MetadataFetcher,
+  type MetadataFetcherServices,
+} from "@/modules/MetadataFetcher";
+import { DOIDiscoveryService } from "@/modules/metadata/DOIDiscoveryService";
+import { MetadataUpdateService } from "@/modules/metadata/MetadataUpdateService";
+import { BookMetadataService } from "@/modules/metadata/BookMetadataService";
 import { createMockItem } from "../../../tests/__mocks__/zotero-items";
+
+function createMockCrossRefAPI() {
+  return {
+    getCrossRefWorkMessage: vi.fn(),
+    fetchWorksByQuery: vi.fn(),
+    getWorkByDOI: vi.fn(),
+    search: vi.fn(),
+    enforceRateLimit: vi.fn(),
+  } as any;
+}
+
+function createMockOpenAlexAPI() {
+  return {
+    getWorkByDOI: vi.fn(),
+    search: vi.fn(),
+    searchExact: vi.fn(),
+    searchOpenAccess: vi.fn(),
+    enforceRateLimit: vi.fn(),
+  } as any;
+}
+
+function createMockSemanticScholarAPI() {
+  return {
+    getPaperByDOI: vi.fn(),
+    search: vi.fn(),
+    searchPapersWithExternalIds: vi.fn(),
+    searchByArxivId: vi.fn(),
+    searchOpenAccess: vi.fn(),
+    enforceRateLimit: vi.fn(),
+  } as any;
+}
 
 describe("Semi-Supervised Learning Paper Metadata Update", () => {
   let fetcher: MetadataFetcher;
+  let mockCrossRefAPI: ReturnType<typeof createMockCrossRefAPI>;
+  let mockMetadataUpdate: MetadataUpdateService;
 
   beforeEach(() => {
+    mockCrossRefAPI = createMockCrossRefAPI();
+    const mockOpenAlexAPI = createMockOpenAlexAPI();
+    const mockSemanticScholarAPI = createMockSemanticScholarAPI();
+    mockMetadataUpdate = new MetadataUpdateService();
+    const mockBookMetadata = new BookMetadataService();
+
+    const doiDiscovery = new DOIDiscoveryService({
+      crossRefAPI: mockCrossRefAPI,
+      openAlexAPI: mockOpenAlexAPI,
+      semanticScholarAPI: mockSemanticScholarAPI,
+    });
+
+    const services: MetadataFetcherServices = {
+      crossRefAPI: mockCrossRefAPI,
+      openAlexAPI: mockOpenAlexAPI,
+      semanticScholarAPI: mockSemanticScholarAPI,
+      doiDiscovery,
+      metadataUpdate: mockMetadataUpdate,
+      bookMetadata: mockBookMetadata,
+    };
+
     fetcher = new MetadataFetcher({
-      config: {
-        downloads: { maxConcurrent: 3 },
-      },
+      config: { downloads: { maxConcurrent: 3 } },
+      services,
     } as never);
 
     vi.stubGlobal("Zotero", {
@@ -27,27 +86,14 @@ describe("Semi-Supervised Learning Paper Metadata Update", () => {
       ItemTypes: {
         getName: vi.fn((typeID: number) => {
           if (typeID === 1) return "journalArticle";
-          if (typeID === 2) return "book";
-          if (typeID === 3) return "conferencePaper";
-          if (typeID === 4) return "preprint";
           return "journalArticle";
         }),
-        getID: vi.fn((name: string) => {
-          if (name === "journalArticle") return 1;
-          if (name === "book") return 2;
-          if (name === "conferencePaper") return 3;
-          if (name === "preprint") return 4;
-          return 1;
-        }),
       },
-      CreatorTypes: {
-        getPrimaryIDForType: vi.fn(() => 1),
-      },
+      CreatorTypes: { getPrimaryIDForType: vi.fn(() => 1) },
       Date: {
-        strToDate: (value: string) => ({
-          year: value.match(/\d{4}/)?.[0],
-        }),
+        strToDate: (value: string) => ({ year: value.match(/\d{4}/)?.[0] }),
       },
+      Translate: { Search: vi.fn() },
     });
   });
 
@@ -63,50 +109,39 @@ describe("Semi-Supervised Learning Paper Metadata Update", () => {
       url: "https://johschool.com/?mod=tammetin&makaleadi=&key=67701",
       extra: "00753 \narXiv: 1406.5298",
       creators: [
-        { firstName: "Abdulselami", lastName: "SARIGÜL" },
+        { firstName: "Abdulselami", lastName: "SARIGL" },
         { firstName: "Shakir", lastName: "Mohamed" },
       ],
     });
 
-    vi.spyOn(fetcher as any, "extractDOI").mockReturnValue(
-      "10.29228/joh.67701",
-    );
-    vi.spyOn(fetcher as any, "fetchDOIMetadataViaTranslator").mockResolvedValue(
-      false,
-    );
-    vi.spyOn(fetcher as any, "fetchCrossRefMetadata").mockResolvedValue({
+    vi.spyOn(fetcher, "extractDOI").mockReturnValue("10.29228/joh.67701");
+    vi.spyOn(fetcher, "fetchDOIMetadataViaTranslator").mockResolvedValue(false);
+    vi.spyOn(
+      fetcher.doiDiscoveryService,
+      "resolvePreferredDoiForMetadata",
+    ).mockResolvedValue("10.29228/joh.67701");
+
+    mockCrossRefAPI.getCrossRefWorkMessage.mockResolvedValue({
       DOI: "10.29228/joh.67701",
       title: ["Semi-Supervised Learning with Deep Generative Models"],
       author: [
-        { given: "Abdulselami", family: "Sarigül" },
+        { given: "Abdulselami", family: "Sarigl" },
         { given: "Shakir", family: "Mohamed" },
       ],
       URL: "https://johschool.com/?mod=tammetin&makaleadi=&key=67701",
       "container-title": ["Journal of Hygiene"],
       published: { "date-parts": [[2023]] },
     });
-    const updateSpy = vi
-      .spyOn(fetcher as any, "updateItemWithMetadata")
-      .mockResolvedValue([
-        "Updated title: Semi-Supervised Learning with Deep Generative Models",
-        "Updated publication title: Journal of Hygiene",
-        "Updated URL: https://johschool.com/?mod=tammetin&makaleadi=&key=67701",
-      ]);
 
-    const result = await (fetcher as any).fetchDOIBasedMetadata(item);
+    vi.spyOn(mockMetadataUpdate, "updateItemWithMetadata").mockResolvedValue([
+      "Updated title: Semi-Supervised Learning with Deep Generative Models",
+      "Updated publication title: Journal of Hygiene",
+    ]);
+    vi.spyOn(mockMetadataUpdate, "supplementDOIMetadata").mockResolvedValue([]);
+
+    const result = await fetcher.fetchDOIBasedMetadata(item);
 
     expect(result.success).toBe(true);
     expect(result.updated).toBe(true);
-    expect(updateSpy).toHaveBeenCalledWith(item, {
-      DOI: "10.29228/joh.67701",
-      title: ["Semi-Supervised Learning with Deep Generative Models"],
-      author: [
-        { given: "Abdulselami", family: "Sarigül" },
-        { given: "Shakir", family: "Mohamed" },
-      ],
-      URL: "https://johschool.com/?mod=tammetin&makaleadi=&key=67701",
-      "container-title": ["Journal of Hygiene"],
-      published: { "date-parts": [[2023]] },
-    });
   });
 });
