+++ src/__tests__/unit/title-only-semi-supervised.test.ts
@@ -1,15 +1,74 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { MetadataFetcher } from "@/modules/MetadataFetcher";
+import {
+  MetadataFetcher,
+  type MetadataFetcherServices,
+} from "@/modules/MetadataFetcher";
+import { DOIDiscoveryService } from "@/modules/metadata/DOIDiscoveryService";
+import { MetadataUpdateService } from "@/modules/metadata/MetadataUpdateService";
+import { BookMetadataService } from "@/modules/metadata/BookMetadataService";
 import { createMockItem } from "../../../tests/__mocks__/zotero-items";
+
+function createMockCrossRefAPI() {
+  return {
+    getCrossRefWorkMessage: vi.fn(),
+    fetchWorksByQuery: vi.fn(),
+    getWorkByDOI: vi.fn(),
+    search: vi.fn(),
+    enforceRateLimit: vi.fn(),
+  } as any;
+}
+
+function createMockOpenAlexAPI() {
+  return {
+    getWorkByDOI: vi.fn(),
+    search: vi.fn(),
+    searchExact: vi.fn(),
+    searchOpenAccess: vi.fn(),
+    enforceRateLimit: vi.fn(),
+  } as any;
+}
+
+function createMockSemanticScholarAPI() {
+  return {
+    getPaperByDOI: vi.fn(),
+    search: vi.fn(),
+    searchPapersWithExternalIds: vi.fn(),
+    searchByArxivId: vi.fn(),
+    searchOpenAccess: vi.fn(),
+    enforceRateLimit: vi.fn(),
+  } as any;
+}
 
 describe("Title-only Semi-Supervised Learning Paper - Existing Test Pattern", () => {
   let fetcher: MetadataFetcher;
+  let mockCrossRefAPI: ReturnType<typeof createMockCrossRefAPI>;
+  let mockMetadataUpdate: MetadataUpdateService;
 
   beforeEach(() => {
+    mockCrossRefAPI = createMockCrossRefAPI();
+    const mockOpenAlexAPI = createMockOpenAlexAPI();
+    const mockSemanticScholarAPI = createMockSemanticScholarAPI();
+    mockMetadataUpdate = new MetadataUpdateService();
+    const mockBookMetadata = new BookMetadataService();
+
+    const doiDiscovery = new DOIDiscoveryService({
+      crossRefAPI: mockCrossRefAPI,
+      openAlexAPI: mockOpenAlexAPI,
+      semanticScholarAPI: mockSemanticScholarAPI,
+    });
+
+    const services: MetadataFetcherServices = {
+      crossRefAPI: mockCrossRefAPI,
+      openAlexAPI: mockOpenAlexAPI,
+      semanticScholarAPI: mockSemanticScholarAPI,
+      doiDiscovery,
+      metadataUpdate: mockMetadataUpdate,
+      bookMetadata: mockBookMetadata,
+    };
+
     fetcher = new MetadataFetcher({
-      config: {
-        downloads: { maxConcurrent: 3 },
-      },
+      config: { downloads: { maxConcurrent: 3 } },
+      services,
     } as never);
 
     vi.stubGlobal("Zotero", {
@@ -27,27 +86,14 @@ describe("Title-only Semi-Supervised Learning Paper - Existing Test Pattern", ()
       ItemTypes: {
         getName: vi.fn((typeID: number) => {
           if (typeID === 1) return "journalArticle";
-          if (typeID === 2) return "book";
-          if (typeID === 3) return "conferencePaper";
-          if (typeID === 4) return "preprint";
           return "journalArticle";
         }),
-        getID: vi.fn((name: string) => {
-          if (name === "journalArticle") return 1;
-          if (name === "book") return 2;
-          if (name === "conferencePaper") return 3;
-          if (name === "preprint") return 4;
-          return 1;
-        }),
       },
-      CreatorTypes: {
-        getPrimaryIDForType: vi.fn(() => 1),
-      },
+      CreatorTypes: { getPrimaryIDForType: vi.fn(() => 1) },
       Date: {
-        strToDate: (value: string) => ({
-          year: value.match(/\d{4}/)?.[0],
-        }),
+        strToDate: (value: string) => ({ year: value.match(/\d{4}/)?.[0] }),
       },
+      Translate: { Search: vi.fn() },
     });
   });
 
@@ -56,25 +102,23 @@ describe("Title-only Semi-Supervised Learning Paper - Existing Test Pattern", ()
     vi.unstubAllGlobals();
   });
 
-  // This test mimics the existing successful test pattern
   it("fetches correct metadata when starting with title only and discovers arXiv DOI", async () => {
     const item = createMockItem({
       title: "Semi-Supervised Learning with Deep Generative Models",
-      // No DOI field initially
       creators: [
         { firstName: "Diederik P.", lastName: "Kingma" },
         { firstName: "Shakir", lastName: "Mohamed" },
       ],
     });
 
-    vi.spyOn(fetcher as any, "extractDOI").mockReturnValue(null);
-    vi.spyOn(fetcher as any, "fetchDOIMetadataViaTranslator").mockResolvedValue(
-      false,
-    );
-    vi.spyOn(fetcher as any, "discoverDOI").mockResolvedValue(
-      "10.48550/arxiv.1406.5298",
-    );
-    vi.spyOn(fetcher as any, "fetchCrossRefMetadata").mockResolvedValue({
+    vi.spyOn(fetcher, "extractDOI").mockReturnValue(null);
+    vi.spyOn(fetcher, "fetchDOIMetadataViaTranslator").mockResolvedValue(false);
+    vi.spyOn(
+      fetcher.doiDiscoveryService,
+      "resolvePreferredDoiForMetadata",
+    ).mockResolvedValue("10.48550/arxiv.1406.5298");
+
+    mockCrossRefAPI.getCrossRefWorkMessage.mockResolvedValue({
       DOI: "10.48550/arxiv.1406.5298",
       title: ["Semi-Supervised Learning with Deep Generative Models"],
       author: [
@@ -87,31 +131,19 @@ describe("Title-only Semi-Supervised Learning Paper - Existing Test Pattern", ()
       "container-title": ["arXiv"],
       published: { "date-parts": [[2014, 6]] },
     });
-    const updateSpy = vi
-      .spyOn(fetcher as any, "updateItemWithMetadata")
-      .mockResolvedValue([
-        "Updated title: Semi-Supervised Learning with Deep Generative Models",
-        "Updated publication title: arXiv",
-        "Updated URL: https://arxiv.org/abs/1406.5298",
-        "Updated DOI: 10.48550/arxiv.1406.5298",
-      ]);
 
-    const result = await (fetcher as any).fetchDOIBasedMetadata(item);
+    vi.spyOn(mockMetadataUpdate, "updateItemWithMetadata").mockResolvedValue([
+      "Updated title: Semi-Supervised Learning with Deep Generative Models",
+      "Updated publication title: arXiv",
+    ]);
+    vi.spyOn(mockMetadataUpdate, "supplementDOIMetadata").mockResolvedValue([]);
+
+    const result = await fetcher.fetchDOIBasedMetadata(item);
 
     expect(result.success).toBe(true);
     expect(result.updated).toBe(true);
-    expect(updateSpy).toHaveBeenCalledWith(item, {
-      DOI: "10.48550/arxiv.1406.5298",
-      title: ["Semi-Supervised Learning with Deep Generative Models"],
-      author: [
-        { given: "Diederik P.", family: "Kingma" },
-        { given: "Danilo J.", family: "Rezende" },
-        { given: "Shakir", family: "Mohamed" },
-        { given: "Max", family: "Welling" },
-      ],
-      URL: "https://arxiv.org/abs/1406.5298",
-      "container-title": ["arXiv"],
-      published: { "date-parts": [[2014, 6]] },
-    });
+    expect(mockCrossRefAPI.getCrossRefWorkMessage).toHaveBeenCalledWith(
+      "10.48550/arxiv.1406.5298",
+    );
   });
 });
