+++ src/modules/metadata/index.ts
@@ -1,0 +1,7 @@
+export { BookMetadataService } from "./BookMetadataService";
+export {
+  DOIDiscoveryService,
+  type DOIDiscoveryServices,
+} from "./DOIDiscoveryService";
+export { MetadataUpdateService } from "./MetadataUpdateService";
+export * from "./types";
