+++ zotero-plugin.config.ts
@@ -13,7 +13,7 @@ export default defineConfig({
     "https://github.com/ydeng11/zotero-zotadata/releases/latest/download/zotadata.xpi",
   version: {
     min: "8.0",
-    max: "8.*",
+    max: "9.*",
   },
   build: {
     assets: ["addon/**/*.*"],
